requirements.txt
torch
torchvision
numpy
opencv-python
gradio
mediapipe
Pillow